avance en las funciones del main parte 2

+++ src/main/kotlin/model/Producto.kt
@@ -4,7 +4,7 @@ abstract class Producto(open val id: Int,
                         open val nombre: String,
                         open var precio: Double,
                         open val categoria: String,
-                        open val tiempoPreparacion: Double) 
+                        open val tiempoPreparacion: Long)
 {
     abstract fun reCalcularPrecio() : Double;
 }
